+++ source/compiler/preprocess.h
@@ -329,7 +329,7 @@ PrGetNextToken (
 void
 PrError (
     UINT8                   Level,
-    UINT8                   MessageId,
+    UINT16                  MessageId,
     UINT32                  Column);
 
 void
